+++ src/main/java/com/codecool/shop/dao/implementation/memory/ProductCategoryDaoMem.java
@@ -1,4 +1,4 @@
-package com.codecool.shop.dao.implementation;
+package com.codecool.shop.dao.implementation.memory;
 
 
 import com.codecool.shop.dao.ProductCategoryDao;
